dio-bank-api/pyproject.toml
[tool.poetry]
name = "dio-bank-api"
version = "0.1.0"
description = ""
authors = ["Rogerio Garcia <rogarfil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "*"
uvicorn = { extras = ["standard"], version = "*" }
databases = { version = "*", extras = ["aiosqlite", "asyncpg"] }
pyjwt = "*"
psycopg2-binary = "*"
pydantic-settings = "*"
alembic = "*"

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
